+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added `HashMap` and `HashSet` implementations to the `v2` version
-
+- Added streaming hashing to the `v2` version
 ## [v0.4.0] - 2024-11-30
 
 - Added `v2` version of the algorithm, available by importing `chibihash::v2::*`. Note that `v2` is missing `StreamingChibiHasher`.